fix: pokemon query hook / ky request / type 서로 분리
